stations from server

+++ Assets/Scripts/WorldmapScripts/StationTierData.cs
@@ -3,5 +3,5 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "StationTierData", menuName = "Technical stuff/StationTierData")]
 public class StationTierData : ScriptableObject
 {
-    public float[] tierReqs;
+    public int[] tierReqs;
 }
